check all possible errors: Service Controller ,ServiceView and ServiceDaoImpl more Junit Tests added to ServiceDaoImplTest

+++ src/tn/medtech/sweng/gemo/daoimpl/ServiceDaoImpl.java
@@ -15,36 +15,41 @@ public class ServiceDaoImpl implements ServiceDao {
         Connection connection= null;
         PreparedStatement preparedStatement=null;
 
-        try{
-            connection= ConnectionConfiguration.getConnection();
-            preparedStatement=connection.prepareStatement("INSERT INTO service (id,name) VALUES(?,?)");
-            preparedStatement.setInt(1,service.getId());
-            preparedStatement.setString(2,service.getName());
-            preparedStatement.executeUpdate();
-            System.out.println();
-            System.out.println();
+        ServiceDaoImpl dao = new ServiceDaoImpl();
 
-            System.out.println("inserted successfully !!");
-        }catch (Exception e){
-            e.printStackTrace();
-        }finally {
-            if(preparedStatement != null){
-                try {
-                    preparedStatement.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
+        if(dao.checkId(service.getId()) >= 0)
+            System.out.println("id already exixts!!");
+        else {
+            try {
+                connection = ConnectionConfiguration.getConnection();
+                preparedStatement = connection.prepareStatement("INSERT INTO service (id,name) VALUES(?,?)");
+                preparedStatement.setInt(1, service.getId());
+                preparedStatement.setString(2, service.getName());
+                preparedStatement.executeUpdate();
+                System.out.println();
+                System.out.println();
+
+                System.out.println("inserted successfully !!");
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                if (preparedStatement != null) {
+                    try {
+                        preparedStatement.close();
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
+                }
+                if (connection != null) {
+                    try {
+                        connection.close();
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
                 }
             }
-            if( connection != null){
-                try {
-                    connection.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
-            }
+
         }
-
-
 
 
 
@@ -55,39 +60,48 @@ public class ServiceDaoImpl implements ServiceDao {
 
         Connection connection=null;
         PreparedStatement preparedStatement=null;
+        ServiceDaoImpl dao = new ServiceDaoImpl();
 
+        if(dao.checkId(id) >= 0) {
 
-        try {
-            connection= ConnectionConfiguration.getConnection();
-            preparedStatement=connection.prepareStatement("UPDATE service SET id=? , name=? WHERE id=?");
-            preparedStatement.setInt(1, service.getId());
-            preparedStatement.setString(2, service.getName() );
-            preparedStatement.setInt(3 , id);
-            preparedStatement.executeUpdate();
-            System.out.println();
-            System.out.println();
-            System.out.println("updated successfully !! ");
-        } catch(Exception e){
-            e.printStackTrace();
-        }finally {
-            if(preparedStatement != null){
+            if (dao.checkId(service.getId()) >= 0  && (service.getId() !=id)) {
+                System.out.println("The id you are trying to insert already exists!");}
+            else {
+
                 try {
-                    preparedStatement.close();
-                } catch (SQLException e) {
+                    connection = ConnectionConfiguration.getConnection();
+                    preparedStatement = connection.prepareStatement("UPDATE service SET id=? , name=? WHERE id=?");
+                    preparedStatement.setInt(1, service.getId());
+                    preparedStatement.setString(2, service.getName());
+                    preparedStatement.setInt(3, id);
+                    preparedStatement.executeUpdate();
+                    System.out.println();
+                    System.out.println();
+                    System.out.println("updated successfully !! ");
+                } catch (Exception e) {
                     e.printStackTrace();
-                }
+                } finally {
+                    if (preparedStatement != null) {
+                        try {
+                            preparedStatement.close();
+                        } catch (SQLException e) {
+                            e.printStackTrace();
+                        }
 
-            }
-            if(connection != null){
-                try {
-                    connection.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
+                    }
+                    if (connection != null) {
+                        try {
+                            connection.close();
+                        } catch (SQLException e) {
+                            e.printStackTrace();
+                        }
+                    }
                 }
             }
+
         }
-
-
+        else
+        { System.out.println("this id does not exist!");}
 
     }
 
@@ -97,38 +111,45 @@ public class ServiceDaoImpl implements ServiceDao {
         Connection conection = null;
         PreparedStatement preparedStatement = null;
 
-        try {
-            //open connection
-            conection = ConnectionConfiguration.getConnection();
-            preparedStatement = conection.prepareStatement("DELETE FROM service WHERE id=?");
-            preparedStatement.setInt(1, id);
-            preparedStatement.executeUpdate();
 
-            //echo
-            System.out.println("DELETE FROM service WHERE name=?");
+        ServiceDaoImpl dao = new ServiceDaoImpl();
 
-        }catch (Exception e){
-            e.printStackTrace();
-        } finally {
-            if(preparedStatement != null){
-                try {
-                    preparedStatement.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
+        if(dao.checkId(id) >= 0) {
+
+            try {
+                //open connection
+                conection = ConnectionConfiguration.getConnection();
+                preparedStatement = conection.prepareStatement("DELETE FROM service WHERE id=?");
+                preparedStatement.setInt(1, id);
+                preparedStatement.executeUpdate();
+
+                //echo
+                System.out.println("Deleted successfully !!");
+
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                if (preparedStatement != null) {
+                    try {
+                        preparedStatement.close();
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
                 }
-            }
 
-            if(conection != null){
-                try {
-                    conection.close();
-                } catch (SQLException e) {
-                    e.printStackTrace();
+                if (conection != null) {
+                    try {
+                        conection.close();
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
                 }
+
+
             }
-
-
         }
-
+        else
+        { System.out.println("this id does not exist!");}
 
 
 
@@ -292,5 +313,52 @@ public class ServiceDaoImpl implements ServiceDao {
 
         return services;
 
+    }
+
+
+    public int  checkId(int id) {
+
+        int r=-1;
+        Connection conection = null;
+        PreparedStatement preparedStatement = null;
+        ResultSet resultSet = null;  // from java sql package ; resultSet is a table of records from your database
+        try {
+            conection = ConnectionConfiguration.getConnection();
+            preparedStatement = conection.prepareStatement("SELECT  id FROM  service WHERE  id = ?");
+            preparedStatement.setInt(1, id);
+            resultSet = preparedStatement.executeQuery();
+
+            while (resultSet.next()) {
+                r=resultSet.getInt("id");
+            }
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            if (resultSet != null) {
+                try {
+                    resultSet.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+
+            if (preparedStatement != null) {
+                try {
+                    preparedStatement.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (conection != null) {
+                try {
+                    conection.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+
+        return r;
     }
 }
